requirements.txt
dash==3.3.0
plotly>=5.22
pandas>=2.2
duckdb>=1.0
pyarrow>=16.0
python-dotenv>=1.0
requests>=2.32
gdown>=5.2
dash-bootstrap-components>=1.5
